refactor: Simplify LPPLTracker and StockLPPLTracker components by utilizing LPPLTrackerBase
- Replaced complex state management and data fetching logic in LPPLTracker and StockLPPLTracker with a new LPPLTrackerBase component.
- Streamlined the fetching of price data and validation of symbols.
- Improved code maintainability and readability by reducing duplication across components.
